Fix component took to long time warnings
These regularly occured, as we were blockingly waiting on a frash SPI
frame update, which could sometimes be too long due to unfortunate
timing. This switches over to a different mechanism where we always
store the latest frame in the core and use that one in ESPHome.

Note that the operation data is now skipped when the frame semaphore is
locked. That is not needed and is an improvement to be made in the
future.