Switch to double frame communication style
After tapping the communications with a logic analyzer, I found the MISO
data to be incorrect: the first part was of a previous frame. After
investigating I found that we are not allowed to modify the buffers
after queueing a transaction. Refactoring this allows us to use the
double frame communication style used in @absalom-muc's repo.