Rework operation data requesting
Communication to the AC wise, similar but this removes the support for
multiple requests simultaneously. That was generating a lot of timeouts
anyway. This also adds a separate layer of timeouts: one that is visible
to the user, hence the extra counter. This hides internal operation data
timeouts by waiting for one minute until communicating it to the user.
Normally, the data is re-requested within that time and thus we don't
have to bother the user with the unknown values.